// src/types/order.ts
export interface ListOrder {
  address: string;
  senderName: string;
  receiverName: string;
  buyerName: string;
  dateReceived: string | null;
  displayStatus: string;
  id: number;
  receiverCountry: string;
  receiverAddress: string;
  receiverDistrict: string;
  japanCode: string;
  jpExportDate: string;
  jpShippingAmount: string | null;
  jpShippingTypeId: string | null;
  mobile: number;
  orderCode: string;
  orderId: string;
  orderCreationDate: string | null;
  orderStatusId: number;
  productName: string | null;
  searchStatus: string;
  sellerName: string;
  shippingCompleteDate: string | null;
  shippingDate: string | null;
  userId: number;
  vnExportDate: string | null;
}